whatsapp-app/backend/requirements.txt
Flask
Flask-SQLAlchemy
requests
flask-cors
python-dotenv